+++ src/main/ts/utils/ImportUtils.ts
@@ -5,3 +5,5 @@ export const isInternalSrcModule = (path: string) => /@(ephox|tiny|tinymce)\/[^\
 export const isInternalPathAlias = (path: string) => /^(ephox|tiny|tinymce)\//.test(path);
 
 export const isMainImport = (path: string) => path.endsWith('/Main');
+
+export const isPublicApiImport = (path: string) => path.endsWith('/PublicApi');